helper collectIds function and check list of ids not to be empty in deleteByIds

+++ nan21.dnet.core.business/src/main/java/net/nan21/dnet/core/business/service/entity/AbstractEntityWriteService.java
@@ -266,7 +266,9 @@ public abstract class AbstractEntityWriteService<E> extends
 	 * @param list
 	 */
 	protected void onDeleteByIds(List<Object> ids) throws BusinessException {
-
+		if (ids == null || ids.size() == 0) {
+			return;
+		}
 		if (IModelWithClientId.class.isAssignableFrom(this.getEntityClass())) {
 			this.em.createQuery(
 					"delete from "
